transaparent weak etag handling via an http client wrapper
